add required-child-reference property
For nodes that have a child that must be a reference.  This allows the
effect system to essentially plan ahead and not choose these nodes
when no reference can be chosen.

It would be even better to have an analysis that detects that a node
must have a reference as a child, but this is a lot easier to
implement, and I'm not going to spend a bunch of time right now on an
analysis that I can obviate with a line of code on the "user" side
that may be needed only in canned-components.

This fixes the bug in commit addbb15 that shows up with seed
1893194732 in the racket-kernel-fuzzer (among others).