xsmith-command-line: add option to force s-exp printing when there are no errors
Because if I have to read a bunch of randomly generated garbage code
to figure out what went wrong, I at least want it to be in a format
that makes it easy to correlate what parts are sub-forms of what.
Reading randomly generated infix/prefix/postfix operator soup is just
terrible.