add debugging help
In particular, add serial numbers to all nodes, and an option to print
them (and other hidden base fields) in the s-expression
representation.

This is primarily for aid in printf-style debugging where a known
program has some issue.  The idea is that I can inspect what's
happening for a particular node by inserting checks for its particular
serial number to print things temporarily while debugging.  I've
wanted this several times, so I figured I should actually put it in.