optimize type checking by adding cached type field to base node
We add a new field to the base node, xsmithcachedtype, and save the
type there once it becomes concrete.  This seems to greatly improve
performance.

However, this commit is not yet correct.  It's generating errors
because there is important side-effectful work that this commit is no
longer running.  So to make this optimization correct it needs to add
in that type checking in another way.  This commit has an attempt to
do that work, but it wasn't quite right.  I'm editing this commit
message while rebasing, so I know the solution.  I just wanted the
commit message to say that while this was a promising start, it wasn't
quite there yet.