overhaul the type caching optimization
This is an ammendment to the type caching optimization that fixes the
side-effect error by changing the strategy a bit.  Rather than caching
only when types become concrete, we cache always.  We also add an API
to re-type a node, which could be manually called but which is
automatically called when a hole is created (well, re-type is a
misnomer here, since it will be typed for the first time), when it is
filled in with `fresh`, when an `edit` is performed (IE the `edit`
property), or... well, I think that's all.

So in this version, as long as a node has been type checked in the
past, the xsmith_type attribute merely accesses the xsmithcachedtype
field.  The _xsmith_type-full function does all the work, including
the side-effecting unification stuff, and is called in the
_xsmith_re-type property.  However, the separate _xsmith_re-type
property exists because when re-typing we actually need to run part of
the full typing code on siblings as well to be sure everything works
out.