Remove now-unused type analysis code
Geeze, a lot of that code was hard to write and hard to get right.
That's a lot of work I'm throwing away.  But maybe the question is why
did I do so much work in a direction that now is so apparently
inferior to what's arguably a more obvious approach?  Life's little
mysteries.