binding-structure: add serial/all
It's like serial, but treats non-binding nodes like binding nodes, in
that each node can only see bindings of nodes before it (in order of
grammar [and list order for list children]).  The motivator for this
addition is for the add-loop-over-container form from
canned-components.  Without this, the collection node (IE the thing to
be looped over) can see the loop variable, which shouldn't be
possible.  (By generation order side-effect it doesn't normally
happen, but if an inner node delays filling using the edit property
the situation can happen.)