remove simple/python -- I don't want to maintain two python fuzzers
Maybe later we can put a simplified version of the python fuzzer back
in, but right now enough things are moving that I'll have to edit both
of them.  Maybe I should have just left it as simple/python, but I
feel like it's graduated past what I indended the “simple” directory
to be.