racket-kernel-fuzzer: fix types of numeric operations
and add duplicates with various types.  Maybe I should not use the
duplicates, but I feel like I want operations for all kinds of
numbers.

The core problem was that I had the type of most arithmetic operators
as (subtype-of number) with arguments of the same type.  But this
breaks with several of them -- eg. nat - nat is not guaranteed to be a
nat, and byte + byte is not guaranteed to be a byte.  I had only paid
attention to this problem with division where it is perhaps more
frequently visible.