sml: adjust printing to put header block inside main
I don't like this, but some MLs that I'm using print out the type of
each value in the program in the output.  I can't figure out how to
turn it off.  I guess Hamlet is not a “serious” SML implementation for
running real programs, but still I feel like it should have a way to
turn that off.

SML/NJ also prints lots of extra crap, but I haven't figured out how
to use it in any case, so I'm less worried about it right now.

But with this adjustment, Hamlet's extra output is always the same, so
it should be easy to filter out with a little wrapper script in the
harness.