canned-components: add single-arg procedures
Previously all the function types handled by canned-components always
had a product-type in the argument position to support functions of
more (or less) than one argument.  But now that I'm implementing a
fuzzer for Standard ML the time for the single argument function has
come.  As it's generic (eg. I'll want this if I implement a Haskell
fuzzer), it's going in canned components.

Also, notably, the single argument function implementation is *much*
simpler than the multi-arg version, and shouldn't need an edit
property to work well.