standardize usage of current-hole macro
I wrote it as a macro that could be used either as an identifier macro
OR as a macro wrapped in parens (a normal macro).  And ever since I've
used it inconsistently, and I've been meaning to clean it up and
standardize it for a long time.  So now it's NOT an identifier macro,
and it MUST be used in parentheses.

This is technically backwards incompatible, but the error message will
point you to the place that's wrong at compile time, so it's a super
easy fix.