add my basic-reproducibility-test script
I wrote it to test when I realized I had broken reproducibility.  I
realized I hadn't checked it in, and it's probably worthwhile.
Notably it does not check for reproducibility between versions of
Racket or anything like that, it just runs a few seeds with whatever
`racket` is on the path and checks a few times that running the same
fuzzer/seed combo makes the same output twice.