canned-components: add depth-weight function
also use it in the various fuzzers so literals are low-weight normally
but become normal weight when at max depth.  Also fix the
choice-weight API so that it can take functions of an argument instead
of thunks (which I had for some reason written that way when I needed
to use it in one place...), and remove the hard-coding of references
as weight 1 in the choice-weight machinery, since the same code is
(better placed) in canned-components.

This gets rid of most lift chains since now the weight disparity
between choosing references and choosing literals essentially flips at
max depth as the right hand side of definitions.