WIP - working on parametric polymorphism
Right now there is a problem.  When generating a polymorphic function,
there is a chance that variable references will be disallowed by the
no-io-conflict test, due to the analysis being over-conservative.
However, the only way to make an expression with parameter-type in a
function body involves variable reference, because the only things of
type alpha are things passed in to the function!

So I probably need to fix the no-io-conflict thing to still allow
reference to at least the parameters of the enclosing functions.

Also, I should probably change the thing that transforms a type into a
polymorphic type to reject the top-level types, so I stop generating
the identity function so much...