reference-options: always allow lifts for functions
We disallow lifting definitions when “any effect” is possible
according to the effect system.  However, this makes it so functions
are not liftable.  In languages with a function literal form this
doesn't keep generation from proceeding, but for languages with only
function definitions it's a problem.  It could be worked around by
adding lambdas and having a lifting pass, but I think we should just
allow functions to be lifted.