canned-components: fix issues with structural record types
I was using a type constraint (the left half of the type-info property
that is used to filter potential choices) that basically had
information from the parent type.  The parent type info is already
propagated through another channel that is in a proper subtype
relationship, but the constraint information is unified (IE
non-subtype symmetrical relationship) that caused issues.  This seems
to fix those issues without causing any new ones.  I don't remember
why I put that code there -- it was clearly a bug, but I don't
remember if there was some reason I tried to put it there.  Hopefully
this doesn't cause any issues.  At any rate, running seeds 1-100 had 2
copies of this error and now there are no (generation) errors, so it
seems like an improvement.  I'll now go run a fuzz campaign and
hopefully this doesn't lead to malformed programs with runtime
errors.  Anyway, I think this fix is correct.