allow Definition node subtypes
I had a rule that lifts could only be done to binder nodes that didn't
have subtypes.  I don't think there was a good reason for this aside
from just knowing which choice object to use for the hole node.  This
commit adds a choice method that returns the ast-node-type that the
generated node will have, which allows me to check rather than
aborting when there is more than one choice.