core: look up binder name and type fields dynamically
This fixes issues with subtyping for binder nodes.  Also fix the issue
where lifting assumed the names 'name and 'type.

Though, frankly, I should have just made 'name and 'type be the
official names for those fields always.  That would have been
smarter.  But since I document that you can choose the names of those
fields, this fixes it to actually live up to that.