add hack to prevent creating parametric types that SML doesn't like
Or more specifically, that fail for most SML implementations.  These
types seem fine to me, and to some SML implementations, but several
barf when they see something like ('a -> 'a -> int).  Since so many
SML implementations fail (not just one), I imagine that maybe support
for such types is optional or something.  At any rate, I'm going to
work around it for now.