add reducible-list-fields property
Some list nodes semantically need a specific number of nodes, such as
function application, tuple constructors, etc.  These should not have
the reducer just delete nodes.  So now there is a property to annotate
which fields are reducible in this way.

The default is to disallow reduction.  However, all definitions in the
canned-components are now explicitly marked, so users could define the
default node as allowing all list nodes to be reduced rather than
marking individually (or to reverse which nodes must be marked).