temporary: comment out rosette requirement in info.rkt
It's only required for the symbolic analysis of cish, which is broken
right now since the Rosette update.  Rosette now requires Racket v8.0,
which is not in the Ubuntu 20.04 image.  Commenting out this
requirement means that cish will fail to build when installing, but we
aren't really using it now anyway.  However, commenting it out enables
me to use other parts of xsmith on rackets older than 8.0, which right
now seems more expedient.

Ultimately, I should either fix the symbolic analysis or stub out the
Rosette imports so it can build without Rosette.